+++ app/src/main/kotlin/de/robinwersich/todue/data/repositories/DatabaseTaskRepository.kt
@@ -7,10 +7,10 @@ import java.time.LocalDate
 class DatabaseTaskRepository(private val taskDao: TaskDao) : TaskRepository {
   override suspend fun insertTask(task: Task) = taskDao.insert(task)
   override suspend fun updateTask(task: Task) = taskDao.update(task)
-  override suspend fun deleteTask(id: Int) = taskDao.delete(id)
-  override suspend fun setText(id: Int, text: String) = taskDao.setText(id, text)
-  override suspend fun setDoneDate(id: Int, doneDate: LocalDate?) =
+  override suspend fun deleteTask(id: Long) = taskDao.delete(id)
+  override suspend fun setText(id: Long, text: String) = taskDao.setText(id, text)
+  override suspend fun setDoneDate(id: Long, doneDate: LocalDate?) =
     taskDao.setDoneDate(id, doneDate)
-  override fun getTask(id: Int) = taskDao.getTask(id)
+  override fun getTask(id: Long) = taskDao.getTask(id)
   override fun getAllTasks() = taskDao.getAllTasks()
 }
